% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{animate_H_from_points}
\alias{animate_H_from_points}
\title{Animate H Function from Points.}
\usage{
animate_H_from_points(
  simulation,
  ROI,
  point_clouds,
  rmax = 200,
  nrval = 20,
  integer_times = TRUE,
  colours = c("red", "black"),
  labels = c("r", "Target", "Actual"),
  time_step = 1,
  time_unit = "ms",
  time_points = NULL
)
}
\arguments{
\item{simulation}{Input simulation.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{point_clouds}{A list of points data, where each data set is given as a matrix or data frame with two columns: x coordinate and y coordinate. Used as target distributions from which spatial statistics are drawn.}

\item{rmax}{Numeric value, the maximum radius for each Ripley's function to be calculated at.}

\item{nrval}{Numeric value, the number of equally-spaced radial values for each Ripley's function to be calculated over.}

\item{integer_times}{Boolean which determines whether times will be rounded to the nearest integer.}

\item{colours}{Colours for target and actual curve respectively.}

\item{labels}{Vector of three characters with first entry representing x axis label, the second representing the name of the target function and the third representing the name of the actual functon.}

\item{time_step}{A numeric value representing the length of one time frame.}

\item{time_unit}{Units of time (optional).}

\item{time_points}{Times at which target function was changed. For discrete time points case only, otherwise leave NULL.}
}
\value{
A gganimate object representing the animated H function and target statistic.
}
\description{
Animates the actual Ripley's H function fitting to other H function(s) from a given point cloud or clouds.
}
