% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{animate_points}
\alias{animate_points}
\title{Animate Point Cloud.}
\usage{
animate_points(
  simulation,
  integer_times = TRUE,
  time_step = 1,
  time_unit = "ms",
  axis_text_size = 12
)
}
\arguments{
\item{simulation}{Input simulation.}

\item{integer_times}{Boolean which determines whether times will be rounded to the nearest integer.}

\item{time_step}{A numeric value representing the length of one time frame.}

\item{time_unit}{Units of time (optional).}

\item{axis_text_size}{Font size of text on axes.}
}
\value{
A gganimate object representing the animated point distribution.
}
\description{
Produces an animation of point dynamics from a simulation.
}
