% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{animate_populations}
\alias{animate_populations}
\title{Animate Multiple Populations.}
\usage{
animate_populations(
  simulation,
  integer_times = TRUE,
  time_step = 1,
  time_unit = "ms",
  labels = NULL,
  text_size = 16,
  axis_text_size = 12,
  point_size = 2,
  legendposition = "right"
)
}
\arguments{
\item{simulation}{Input simulation.}

\item{integer_times}{Boolean which determines whether times will be rounded to the nearest integer.}

\item{time_step}{A numeric value representing the length of one time frame.}

\item{time_unit}{Units of time (optional).}

\item{labels}{A vector of characters representing the name of each population type. Will be displayed in the animation legend if not NULL.}

\item{text_size}{Text size of labels in legend.}

\item{axis_text_size}{Font size of text on axes.}

\item{point_size}{Size of each point in the simulation.}
}
\value{
A gganimate object representing the animated point distribution, including multiple populations.
}
\description{
Produces an animation of population dynamics from a simulation.
}
