% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{animate_target}
\alias{animate_target}
\title{Animate Target Function.}
\usage{
animate_target(
  simulation,
  ROI,
  target,
  rmax = 200,
  nrval = 20,
  integer_times = TRUE,
  colours = c(rgb(212/255, 30/255, 125/255), rgb(245/255, 93/255, 0/255)),
  labels = c("r", "K(r)"),
  time_step = 1,
  time_unit = "ms",
  axis_text_size = 16,
  title_text_offset = 4,
  legendposition = "right"
)
}
\arguments{
\item{simulation}{Input simulation.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{target}{The target curve.}

\item{rmax}{Numeric value, the maximum radius for each K function to be calculated at.}

\item{nrval}{Numeric value, the number of equally-spaced radial values for each K function to be calculated over.}

\item{integer_times}{Boolean which determines whether times will be rounded to the nearest integer.}

\item{colours}{Colours for target and actual curve respectively.}

\item{labels}{Vector of characters representing axis labels.}

\item{time_step}{A numeric value representing the length of one time frame.}

\item{time_unit}{Units of time (optional).}

\item{axis_text_size}{Font size of text on axes.}

\item{title_text_offset}{Increment in font size for axis titles.}
}
\value{
A gganimate object representing the animated K function and target statistic.
}
\description{
Animates the actual Ripley's K function of the point distribution in the simulation, compared to a target curve.
}
