% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{csrwithremoval}
\alias{csrwithremoval}
\title{Completely Spatially Random Simulator with Point Removal.}
\usage{
csrwithremoval(
  step_size = 50,
  ROI = 1000,
  times = 100,
  populations = 3,
  initial_numbers = 1000,
  populations_to_remove = 1,
  removal_rates = 0.1,
  dimensions = 2
)
}
\arguments{
\item{step_size}{Fixed step size.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{times}{Number of time steps or iterations of the simulation.}

\item{populations}{Number of distinct point populations to consider.}

\item{initial_numbers}{Initial number of points in each population. Can specify for each population with a vector.}

\item{populations_to_remove}{A vector of population numbers. Points from these populations will be removed at their corresponding removal rates.}

\item{removal_rates}{A vector of removal rates, determines what random proportion of each population will be removed in each frame.}

\item{dimensions}{Dimensions of data set.}
}
\value{
A molecular dynamics simulation. This takes the form of a list, containing a point cloud for each time frame. Will also contain spatial statistics at each time frame if output_target is TRUE.
}
\description{
Simulates CSR distributions moving with fixed random step size and removes points of a certain population at a given rate.
}
