% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{default_aggregate}
\alias{default_aggregate}
\title{Default Aggregate.}
\usage{
default_aggregate(target)
}
\arguments{
\item{target}{List of targets, as output by target functions.}
}
\value{
List of targets, as output by target functions.
}
\description{
Does nothing, returns targets as list. Typically used for debugging.
}
