% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{default_correction}
\alias{default_correction}
\title{Default Correction}
\usage{
default_correction(points, ROI)
}
\arguments{
\item{points}{A matrix or data frame with two columns: x and y coordinates of each point.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}
}
\value{
New point coordinates after correction.
}
\description{
Does not correct points which escape boundary.
}
