% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{default_velocity_2D}
\alias{default_velocity_2D}
\title{Default Velocity.}
\usage{
default_velocity_2D(D_min, D_max, ROI, error)
}
\arguments{
\item{D_min}{Minimum step size.}

\item{D_max}{Maximum step size.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{error}{A vector of error values, as output by error function.}
}
\value{
Offsets for each point (in this case, 0).
}
\description{
No movement, returns 0.
}
