% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{distripleysim}
\alias{distripleysim}
\title{Ripley's Simulator with Distinct Time Points}
\usage{
distripleysim(
  D_min = 0,
  D_max = 50,
  ROI = 1000,
  time_points = 1000,
  point_clouds = NULL,
  rmax = 200,
  nrval = 20,
  dimensions = 2,
  targets = NULL,
  number_of_points = NULL,
  initial_distribution = NULL
)
}
\arguments{
\item{D_min}{Minimum step size.}

\item{D_max}{Maximum step size.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{time_points}{A vector of time points in the simulation. Each time point represents a change in target function. The final value in this vector is the total simulation time.}

\item{rmax}{Numeric value, the maximum radius for each K function to be calculated at.}

\item{nrval}{Numeric value, the number of equally-spaced radial values for each K function to be calculated over.}

\item{dimensions}{Dimensions of data set.}

\item{targets}{Manual input for target function. Can be left as NULL as long as a target point cloud is provided. Length of each target must be equal to nrval.}

\item{number_of_points}{Number of points to use in simulation.}

\item{initial_distribution}{The initial frame of the simulation, used to specify pre-determined spatial organisation. If left NULL, a uniformly random distribution will be used.}

\item{point_cloud}{A matrix or data frame with two columns: x and y coordinates of each point. Used as a target distribution from which spatial statistics are drawn.}

\item{output_target}{Boolean, determines whether spatial statistics will be output alongside point cloud at each frame. Useful for tracking changes in spatial statistics.}
}
\value{
A molecular dynamics simulation. This takes the form of a list, containing a point cloud for each time frame. Will also contain spatial statistics at each time frame if output_target is TRUE.
}
\description{
Iteratively fits each point to the global Ripley's K function, target function changes at distinct time points.
}
