% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{HtoK}
\alias{HtoK}
\title{H to K}
\usage{
HtoK(H)
}
\arguments{
\item{H}{A matrix or data frame with two columns: r and H.}
}
\value{
The global K function derived from H.
}
\description{
Converts Ripley's H function back to a K function.
}
