% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{KtoH}
\alias{KtoH}
\title{K to H}
\usage{
KtoH(K)
}
\arguments{
\item{K}{A matrix or data frame with two columns: r and K.}
}
\value{
The global H function derived from K.
}
\description{
Converts Ripley's K function to a H function.
}
