% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{plotlines}
\alias{plotlines}
\title{Plot Lines.}
\usage{
plotlines(
  data,
  labels = c("x", "y"),
  xlimits = NULL,
  ylimits = NULL,
  colours = NULL,
  textsize = 20,
  titletextoffset = 4,
  removetitles = FALSE,
  linewidth = 1,
  xpand = NULL,
  ypand = c(0, 0),
  xpandadd = c(0, 0),
  ypandadd = c(0, 0),
  xangle = 0,
  xhjust = 0.5,
  xvjust = 0,
  borderthickness = 1,
  xpandmultiplier = 0.01,
  linetype = "solid",
  background = "transparent"
)
}
\arguments{
\item{data}{A matrix or data frame. First column represents x coordinates. Each additional column represents y coordinates for a new line.}

\item{labels}{A vector of characters representing the x and y axis labels.}

\item{xlimits}{Vector with lower and upper bound in x direction.}

\item{ylimits}{Vector with lower and upper bound in y direction.}

\item{colours}{Vector of RGB colours representing colours to be used in case of multiple lines. Chosen automatically from a colourblind-friendly palette if not given.}

\item{textsize}{Numeric, font size of axis text size.}

\item{titletextoffset}{Numeric, increment in font size for titles compared to axis text.}

\item{removetitles}{Boolean, determines whether axis text is removed.}

\item{linewidth}{Numeric, thickness of lines in plot.}

\item{xpand}{Vector of scalar multipliers for additional space (expand) given in x direction.}

\item{ypand}{Vector of scalar multipliers for additional space (expand) given in y direction.}

\item{xpandadd}{Vector of numeric increments for additional space (expand) given in x direction.}

\item{ypandadd}{Vector of numeric increments for additional space (expand) given in y direction.}

\item{xangle}{Numeric, angle of x axis text.}

\item{xhjust}{Numeric, horizontal displacement of x axis text.}

\item{xvjust}{Numeric, vertical displacement of x axis text.}

\item{borderthickness}{Numeric, line thickness of plot border.}

\item{xpandmultiplier}{Numeric for automatically scaling xpand if not given.}

\item{linetype}{Linetype, see ggplot2 line types.}

\item{background}{Character or RGB colour. Background colour of plot.}
}
\value{
A stylised ggplot of the lines.
}
\description{
Creates a nice plot of some lines.
}
