% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{plotpoints}
\alias{plotpoints}
\title{Plot Point Cloud.}
\usage{
plotpoints(
  points,
  labels = c("x", "y"),
  textsize = 16,
  titletextoffset = 0,
  removetitles = TRUE,
  pointsize = 1.25,
  borderthickness = 1,
  colours = NULL,
  repeatcolours = TRUE,
  background = "transparent",
  xlimits = NULL,
  ylimits = NULL
)
}
\arguments{
\item{points}{A matrix or data frame with two columns: x and y coordinates of each point. An additional third column for population type may also be given, this will be coloured automatically.}

\item{labels}{A vector of characters representing the x and y axis labels.}

\item{textsize}{Numeric, font size of axis text size.}

\item{titletextoffset}{Numeric, increment in font size for titles compared to axis text.}

\item{removetitles}{Boolean, determines whether axis text is removed.}

\item{pointsize}{Numeric, size of points in point cloud plot.}

\item{borderthickness}{Numeric, line thickness of plot border.}

\item{colours}{Vector of RGB colours representing colours to be used in case of multiple populations. Chosen automatically from a colourblind-friendly palette if not given.}

\item{repeatcolours}{Boolean, allows plot to repeat colours as many times as necessary to colour all populations. If FALSE, you will need to specify a number of colours equal to the number of different populations.}

\item{background}{Character or RGB colour. Background colour of plot.}

\item{xlimits}{Vector with lower and upper bound in x direction.}

\item{ylimits}{Vector with lower and upper bound in y direction.}
}
\value{
A stylised ggplot of the point cloud.
}
\description{
Creates a nice plot of any input point cloud.
}
