% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{pointwise_ratiometric_error_main}
\alias{pointwise_ratiometric_error_main}
\title{Pointwise Ratiometric Error Main.}
\usage{
pointwise_ratiometric_error_main(target, actual, n)
}
\arguments{
\item{target}{A list of targets, as output by target functions.}

\item{actual}{A list of actual point statistics, taken points input to target functions.}

\item{n}{Number of points, the length of actual.}
}
\value{
A vector of error values, one for each point.
}
\description{
Calculates average ratiometric error of each point across a 2D curve, slightly faster. Compares each point to a distinct local target. Used in simulator function.
}
