% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{pointwise_ratiometric_error_main_global}
\alias{pointwise_ratiometric_error_main_global}
\title{Pointwise Ratiometric Error Main Global.}
\usage{
pointwise_ratiometric_error_main_global(target, actual, n, num)
}
\arguments{
\item{target}{A target, aggregated from output of a target function.}

\item{actual}{A list of actual point statistics, taken points input to target functions.}

\item{n}{Number of points, the length of actual.}

\item{num}{Placeholder argument, currently unused.}
}
\value{
A vector of error values, one for each point.
}
\description{
Calculates average ratiometric error of each point across a 2D curve. Used in simulator function.
}
