% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{pointwise_ratiometric_error_pop}
\alias{pointwise_ratiometric_error_pop}
\title{Pointwise Ratiometric Error Multiple Populations}
\usage{
pointwise_ratiometric_error_pop(targets, actuals, n)
}
\arguments{
\item{targets}{A list of targets, as output by target functions. One for each population.}

\item{actuals}{A list of actual point statistics, taken points input to target functions.}

\item{n}{Number of points, the length of actual.}
}
\value{
A vector of error values, one for each point.
}
\description{
Calculates average ratiometric error of each point across a 2D curve for each point population.
}
