% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{population_changer_with_removal}
\alias{population_changer_with_removal}
\title{Population Changer}
\usage{
population_changer_with_removal(simulation, A, B, C, D)
}
\arguments{
\item{simulation}{Input simulation.}

\item{A}{A vector of population numbers to change.}

\item{B}{A vector of population numbers to change A into.}

\item{C}{A vector of population numbers which bring about change from A to B.}

\item{D}{A vector of distances for which C brings about change.}
}
\value{
A molecular dynamics simulation, with population interactions now recorded.
}
\description{
Changes points from population A to population B when they come within a distance D of population C. Accounts for removed points. Used to post-process multiple population simulations.
}
