% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{preferential_ratiometric_error_all}
\alias{preferential_ratiometric_error_all}
\title{Preferential Ratiometric Error All}
\usage{
preferential_ratiometric_error_all(targets, actual, n)
}
\arguments{
\item{targets}{A list of targets, as output by target functions.}

\item{actual}{A list of actual point statistics, taken points input to target functions.}

\item{n}{Number of points, the length of actual.}
}
\value{
A vector of error values, one for each point.
}
\description{
Calculates the ratiometric error of each point across multiple 2D curves. Returns all errors for tracking changes.
}
