% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{preferential_ratiometric_error_all_from_zero}
\alias{preferential_ratiometric_error_all_from_zero}
\title{Preferential Ratiometric Error All from Zero}
\usage{
preferential_ratiometric_error_all_from_zero(targets, actual, n)
}
\arguments{
\item{targets}{A list of targets, as output by target functions.}

\item{actual}{A list of actual point statistics, taken points input to target functions.}

\item{n}{Number of points, the length of actual.}
}
\value{
A vector of error values, one for each point.
}
\description{
Calculates the ratiometric error of each point across multiple 2D curves. Returns all errors for tracking changes. Baseline at 0, useful for CSR targets.
}
