% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{reflection_correction}
\alias{reflection_correction}
\title{Reflection Correction}
\usage{
reflection_correction(points, ROI)
}
\arguments{
\item{points}{A matrix or data frame with two columns: x and y coordinates of each point.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}
}
\value{
New point coordinates after correction.
}
\description{
Causes points to rebound off of sides of ROI.
}
