% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{reflection_correction_pop}
\alias{reflection_correction_pop}
\title{Reflection Correction Multiple Populations.}
\usage{
reflection_correction_pop(clouds, ROI)
}
\arguments{
\item{clouds}{A list of points data, where each data set is given as a matrix or data frame with two columns: x and y coordinates of each point.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}
}
\value{
New point coordinates after correction.
}
\description{
Causes points to rebound off of sides of ROI. Used in multiple populations cases.
}
