% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripley_aggregate}
\alias{ripley_aggregate}
\title{Ripley Aggregate.}
\usage{
ripley_aggregate(target)
}
\arguments{
\item{target}{List of targets, as output by target functions.}
}
\value{
Global K function derived from mean of all targets.
}
\description{
Takes average of second column, converts list of localised Ripley's K functions into global K function.
}
