% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripley_aggregate_pop}
\alias{ripley_aggregate_pop}
\title{Ripley Aggregate for Multiple Populations}
\usage{
ripley_aggregate_pop(targets)
}
\arguments{
\item{target}{List of targets, as output by target functions.}
}
\value{
List of global K functions, one for each population.
}
\description{
Takes average of second column for all targets. Used for multiple populations cases.
}
