% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripley_H_aggregate}
\alias{ripley_H_aggregate}
\title{Ripley H Aggregate.}
\usage{
ripley_H_aggregate(target)
}
\arguments{
\item{target}{List of targets, as output by target functions.}
}
\value{
Global H function over all points.
}
\description{
Converts list of localised Ripley's K functions into global H function.
}
