% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripley_H_aggregate_pop}
\alias{ripley_H_aggregate_pop}
\title{Ripley H Aggregate for Multiple Populations}
\usage{
ripley_H_aggregate_pop(targets)
}
\arguments{
\item{target}{List of targets, as output by target functions.}
}
\value{
List of global H functions, one for each population.
}
\description{
Calculates global H function for for all targets. Used for multiple populations cases.
}
