% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripley_target}
\alias{ripley_target}
\title{Local Ripley's Target.}
\usage{
ripley_target(points, ROI, rmax = 200, nrval = 20)
}
\arguments{
\item{points}{A matrix or data frame with two columns: x and y coordinates of each point.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{rmax}{Numeric value, the maximum radius for each K function to be calculated at.}

\item{nrval}{Numeric value, the number of equally-spaced radial values for each K function to be calculated over.}
}
\value{
A list of the local K functions, one for each row (point) in points.
}
\description{
Calculates the localised, linearised Ripley's K function for each point.
}
