% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripley_target_euclidean_main}
\alias{ripley_target_euclidean_main}
\title{Main Euclidean Local Ripley's Target}
\usage{
ripley_target_euclidean_main(points, ROI, R, lambda, number_of_points)
}
\arguments{
\item{points}{A matrix or data frame with two columns: x and y coordinates of each point.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{R}{A vector of radial values for which to calculate the K function over.}

\item{lambda}{Numeric value, used to weight the K function. Calculated automatically, should not be altered.}

\item{number_of_points}{Numeric value, the number of data points. Calculated automatically if not given.}
}
\value{
A list of the local K functions, one for each row (point) in points.
}
\description{
Calculates the localised, linearised Ripley's K function for each point using Euclidean distance, slightly faster. Uses alternative calculation for computational efficiency. Used in main simulator.
}
