% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{ripleyk}
\alias{ripleyk}
\title{Ripley's K Function}
\usage{
ripleyk(points, ROI, rmax = 200, nrval = 20)
}
\arguments{
\item{points}{A matrix or data frame with two columns: x and y coordinates of each point.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{rmax}{Numeric value, the maximum radius for each K function to be calculated at.}

\item{nrval}{Numeric value, the number of equally-spaced radial values for each K function to be calculated over.}
}
\value{
The global K function for the points data set.
}
\description{
Calculate the global Ripley's K function for a point cloud.
}
