% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{simulatecircularclusters}
\alias{simulatecircularclusters}
\title{Simulate Circular Clusters.}
\usage{
simulatecircularclusters(
  ROI = 1000,
  number_of_clusters = 10,
  cluster_radius = 30,
  points_per_cluster = 20,
  background_to_cluster_ratio = 0.33
)
}
\arguments{
\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{number_of_clusters}{Integer, number of circular clusters to generate.}

\item{cluster_radius}{Numeric, radius of the clusters.}

\item{points_per_cluster}{Integer, the number of points in each cluster.}

\item{background_to_cluster_ratio}{Numeric value, the ratio of background points to cluster points. Rounded to nearest integer.}
}
\value{
A matrix with two columns: the x and y coordinates of the point cloud. This point cloud will contain circular clusters.
}
\description{
Simulates a static point cloud with circular clusters.
}
