% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{simulatecsr}
\alias{simulatecsr}
\title{Simulate Complete Spatially Random Clusters.}
\usage{
simulatecsr(ROI, number_of_points)
}
\arguments{
\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{number_of_points}{Integer, number of points in point cloud.}
}
\value{
A matrix with two columns: the x and y coordinates of the point cloud. This point cloud will contain randomly placed points only.
}
\description{
Simulates a static point cloud with randomly placed points.
}
