% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{squeezereleaseripleysim}
\alias{squeezereleaseripleysim}
\title{Simultaneous Targets with Squeeze and Release}
\usage{
squeezereleaseripleysim(
  D_min = 0,
  D_max = 50,
  ROI = 1000,
  times = 1000,
  point_clouds = NULL,
  rmax = 200,
  nrval = 20,
  dimensions = 2,
  targets = NULL,
  number_of_points = NULL,
  initial_distribution = NULL,
  output_target = TRUE,
  perturbation_time = 500,
  number_cross_linked_clusters = 1,
  number_cross_linked = 10,
  cross_linked_cluster_radius = 5
)
}
\arguments{
\item{D_min}{Minimum step size.}

\item{D_max}{Maximum step size.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{times}{Number of time steps or iterations of the simulation.}

\item{point_clouds}{A list of points data, where each data set is given as a matrix or data frame with two columns: x and y coordinates of each point. Used as target distributions from which spatial statistics are drawn.}

\item{rmax}{Numeric value, the maximum radius for each K function to be calculated at.}

\item{nrval}{Numeric value, the number of equally-spaced radial values for each K function to be calculated over.}

\item{dimensions}{Dimensions of data set.}

\item{targets}{Manual input for target functions. Can be left as NULL as long as target point clouds are provided.}

\item{number_of_points}{Number of points to use in simulation.}

\item{initial_distribution}{The initial frame of the simulation, used to specify pre-determined spatial organisation. If left NULL, a uniformly random distribution will be used.}

\item{output_target}{Boolean, determines whether spatial statistics will be output alongside point cloud at each frame. Useful for tracking changes in spatial statistics.}

\item{perturbation_time}{The time frame at which the perturbation occurs. Leave as 0 for no perturbation.}

\item{number_cross_linked_clusters}{Number of clusters to be created.}

\item{number_cross_linked}{Number of points in each induced cluster.}

\item{cross_linked_cluster_radius}{Radius of induced clusters.}
}
\value{
A molecular dynamics simulation. This takes the form of a list, containing a point cloud for each time frame. Will also contain spatial statistics at each time frame if output_target is TRUE.
}
\description{
Iteratively fits each point to one of multiple possible Ripley's functions (depending on which is closest), squeezes points together at a given time then releases.
}
