% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{step_velocity}
\alias{step_velocity}
\title{Step Velocity.}
\usage{
step_velocity(D_min, D_max, ROI, error, threshold = 0.5)
}
\arguments{
\item{D_min}{Minimum step size.}

\item{D_max}{Maximum step size.}

\item{ROI}{A vector of two values representing the width and height of the region of interest, respectively.}

\item{error}{A vector of error values, as output by error function.}

\item{threshold}{A numeric value. If a point's error is above the threshold, it will be offset with step size D_max, otherwise it will be offset with step size D_min.}
}
\value{
Offsets for each point.
}
\description{
Determines velocity at either minimum or maximum, depending on whether error surpasses a given threshold.
}
