% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolecularDynamicsFunctions.R
\name{track_changes}
\alias{track_changes}
\title{Track Changes}
\usage{
track_changes(simulation, A, time_step = 1)
}
\arguments{
\item{simulation}{Input simulation.}

\item{A}{A population number to track.}

\item{time_step}{A numeric value representing the length of one time frame.}
}
\value{
A molecular dynamics simulation, with population interactions now recorded.
}
\description{
Tracks changes in number of points of population A over time.
}
